% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromfx.R
\name{getAtlasPeaks}
\alias{getAtlasPeaks}
\title{getAtlasPeaks}
\usage{
getAtlasPeaks(peaksummits)
}
\arguments{
\item{peaksummits}{GRanges of peak summits with coloumn "score"}
}
\value{
GRanges
}
\description{
function that takes a GRanges of peak summits and geenerates fixed width peak atlas
}
\author{
Ashley S Doane
}
\keyword{straw}
