% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromfx.R
\name{readNarrowPeakSummits}
\alias{readNarrowPeakSummits}
\title{readNarrowPeakSummits}
\usage{
readNarrowPeakSummits(npeaks, genome = "hg38")
}
\arguments{
\item{npeaks}{paths to sample narrowPeak file}

\item{genome}{reference genome, either "hg38 (default) or "mm10"}
}
\value{
GRanges
}
\description{
function that reads narrowPeak ATAC peaks and returns 500bp ranges centered on summit
}
\keyword{peaks}
