% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromfx.R
\name{readPeakSummits}
\alias{readPeakSummits}
\title{readPeakSummits}
\usage{
readPeakSummits(psum, genome = "hg38")
}
\arguments{
\item{psum}{paths to sample peak summits bed file}

\item{genome}{reference genome, either "hg38 (default) or "mm10"}
}
\value{
GRanges
}
\description{
function that reads ATAC peaks summits
}
\keyword{peaks}
